Repository Files/requirements.txt
streamlit==1.25.0
psycopg2-binary==2.9.1
feedparser==6.0.8
pandas==1.3.4
numpy==1.21.2
requests==2.26.0
protobuf==3.20.3
